fixup! Do not expect CWD to be passed for config
